Reborn: Fix signed/unsigned and type-casting issues (2)
This change moves the NGMP online-services cleanup into OnlineServices_Init.cpp and fixes signed/unsigned comparisons and casts across networking, AI, UI, and gameplay code. The updates avoid out-of-range/index bugs, preserve legacy behavior, and address compiler warnings that surfaced in newer builds.

+++ GeneralsMD/Code/GameEngine/Include/GameNetwork/GeneralsOnline/OnlineServices_MatchmakingInterface.h
@@ -52,7 +52,7 @@ public:
 
 	PlaylistEntry GetCachedPlaylistFromIndex(int index) const
 	{
-		if (index < m_vecCachedPlaylists.size())
+		if ((size_t)index < m_vecCachedPlaylists.size())
 		{
 			return m_vecCachedPlaylists.at(index);
 		}
